+++ apps/meteor/client/views/omnichannel/customFields/CustomFieldsForm.stories.tsx
@@ -1,13 +1,12 @@
 import { Box } from '@rocket.chat/fuselage';
-import { action } from '@storybook/addon-actions';
 import type { ComponentMeta, ComponentStory } from '@storybook/react';
 import React from 'react';
 
-import NewCustomFieldsForm from './NewCustomFieldsForm';
+import EditCustomFields from './EditCustomFields';
 
 export default {
-	title: 'Omnichannel/NewCustomFieldsForm',
-	component: NewCustomFieldsForm,
+	title: 'Omnichannel/CustomFields',
+	component: EditCustomFields,
 	decorators: [
 		(fn) => (
 			<Box maxWidth='x600' alignSelf='center' w='full' m={24}>
@@ -15,23 +14,7 @@ export default {
 			</Box>
 		),
 	],
-} as ComponentMeta<typeof NewCustomFieldsForm>;
+} as ComponentMeta<typeof EditCustomFields>;
 
-export const Default: ComponentStory<typeof NewCustomFieldsForm> = (args) => <NewCustomFieldsForm {...args} />;
+export const Default: ComponentStory<typeof EditCustomFields> = (args) => <EditCustomFields {...args} />;
 Default.storyName = 'CustomFieldsForm';
-Default.args = {
-	values: {
-		field: '',
-		label: '',
-		scope: 'visitor',
-		visibility: true,
-		regexp: '',
-	},
-	handlers: {
-		handleField: action('handleField'),
-		handleLabel: action('handleLabel'),
-		handleScope: action('handleScope'),
-		handleVisibility: action('handleVisibility'),
-		handleRegexp: action('handleRegexp'),
-	},
-};