+++ ee/packages/license/src/definition/ILicenseV3.ts
@@ -8,7 +8,7 @@ export interface ILicenseV3 {
 	information: {
 		id?: string;
 		autoRenew: boolean;
-		visualExpiration: Timestamp;
+		visualExpiration?: Timestamp;
 		notifyAdminsAt?: Timestamp;
 		notifyUsersAt?: Timestamp;
 		trial: boolean;