+++ packages/web-ui-registration/src/components/LoginSwitchLanguageFooter.tsx
@@ -1,20 +1,10 @@
 import { Button } from '@rocket.chat/fuselage';
 import { useLocalStorage } from '@rocket.chat/fuselage-hooks';
 import { HorizontalWizardLayoutCaption } from '@rocket.chat/layout';
+import { normalizeLanguage } from '@rocket.chat/tools';
 import { type TranslationLanguage, useSetting, useLoadLanguage, useLanguage, useLanguages } from '@rocket.chat/ui-contexts';
 import { type ReactElement, type UIEvent, useMemo, useEffect } from 'react';
 import { Trans, useTranslation } from 'react-i18next';
-
-const normalizeLanguage = (language: string): string => {
-	// Fix browsers having all-lowercase language settings eg. pt-br, en-us
-	const regex = /([a-z]{2,3})-([a-z]{2,4})/;
-	const matches = regex.exec(language);
-	if (matches) {
-		return `${matches[1]}-${matches[2].toUpperCase()}`;
-	}
-
-	return language;
-};
 
 const useSuggestedLanguages = ({
 	browserLanguage = normalizeLanguage(window.navigator.language ?? 'en'),